refactor: information, symptomType 컴포넌트 분리
