+++ src/components/diagnosisPage/AnswerButtons.tsx
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 import { IAnswerButtonProps } from "../../interfaces/diagnosisPage";
-import RoundButton from "../buttons/RoundButton";
+import RoundButton from "../roundButton";
 import theme from "../../lib/theme";
 import { Body_1 } from "../../lib/fontStyle";
 import { useEffect } from "react";
@@ -20,8 +20,7 @@ const AnswersContainer = styled.section<{ ansCount: number }>`
   flex-direction: column;
   align-items: center;
 
-  margin-top: ${({ ansCount }) =>
-    ansCount === 2 ? 13.4 : ansCount === 3 ? 10.2 : 7}rem;
+  margin-top: ${({ ansCount }) => (ansCount === 2 ? 13.4 : ansCount === 3 ? 10.2 : 7)}rem;
 `;
 const ButtonBox = styled.section`
   width: calc(var(--vw, 1vw) * 100 - 4rem);
@@ -36,13 +35,10 @@ const Button = styled.section<{ selected: boolean }>`
 
   padding: ${({ selected }) => (selected ? "1.4rem 2.3rem" : "1.3rem 2.3rem")};
 
-  background-color: ${({ selected, theme }) =>
-    selected ? theme.color.sub_blue : "transparent"};
-  color: ${({ selected, theme }) =>
-    selected ? theme.color.blue : theme.color.grey_300};
+  background-color: ${({ selected, theme }) => (selected ? theme.color.sub_blue : "transparent")};
+  color: ${({ selected, theme }) => (selected ? theme.color.blue : theme.color.grey_300)};
 
-  border: ${({ selected, theme }) =>
-    !selected && `0.1rem solid ${theme.color.grey_650}`};
+  border: ${({ selected, theme }) => !selected && `0.1rem solid ${theme.color.grey_650}`};
   border-radius: 9rem;
 
   cursor: pointer;
@@ -68,12 +64,7 @@ const NextButton = styled.section`
   background: linear-gradient(180deg, rgba(31, 37, 79, 0) 0%, #23284b 50%);
 `;
 
-const AnswerButtons = ({
-  question,
-  selectedAnswer,
-  setSelectedAnswer,
-  handleNext,
-}: IAnswerButtonProps) => {
+const AnswerButtons = ({ question, selectedAnswer, setSelectedAnswer, handleNext }: IAnswerButtonProps) => {
   const answers = question.answers;
   const isMultiple = question.is_multiple;
 
@@ -143,18 +134,11 @@ const AnswerButtons = ({
         <NextButton onClick={handleMultipleAnswer}>
           <RoundButton
             outline="none"
-            backgroundColor={
-              selectedAnswer.length === 0
-                ? theme.color.grey_650
-                : theme.color.blue
-            }
-            color={
-              selectedAnswer.length === 0
-                ? theme.color.grey_400
-                : theme.color.grey_100
-            }
-            text="다음 단계"
-          />
+            backgroundColor={selectedAnswer.length === 0 ? theme.color.grey_650 : theme.color.blue}
+            color={selectedAnswer.length === 0 ? theme.color.grey_400 : theme.color.grey_100}
+          >
+            다음 단계
+          </RoundButton>
         </NextButton>
       )}
     </Container>
