+++ src/components/mainPage/BottomButtons.tsx
@@ -1,7 +1,7 @@
 import { Dispatch } from "react";
 import styled from "styled-components";
 import theme from "../../lib/theme";
-import RoundButton from "../buttons/RoundButton";
+import RoundButton from "../roundButton";
 import { useNavigate } from "react-router-dom";
 
 const Buttons = styled.section`
@@ -29,14 +29,18 @@ const BottomButtons = ({ setModal }: { setModal: Dispatch<boolean> }) => {
   return (
     <Buttons>
       <ButtonBox onClick={() => navigate("/info")}>
-        <RoundButton outline="none" backgroundColor={theme.color.green} color={theme.color.grey_800} text={"빠른 진단 시작하기"} />
+        <RoundButton outline="none" backgroundColor={theme.color.green} color={theme.color.grey_800}>
+          빠른 진단 시작하기
+        </RoundButton>
       </ButtonBox>
       <ButtonBox
         onClick={() => {
           setModal(true);
         }}
       >
-        <RoundButton outline="none" backgroundColor={theme.color.blue} color={theme.color.grey_100} text={"나의 진단기록장 보기"} />
+        <RoundButton outline="none" backgroundColor={theme.color.blue} color={theme.color.grey_100}>
+          나의 진단기록장 보기
+        </RoundButton>
       </ButtonBox>
     </Buttons>
   );
