+++ src/pages/information/gender/index.style.tsx
@@ -1,0 +1,24 @@
+import styled from "styled-components";
+import { Body_4 } from "../../../lib/fontStyle";
+
+export const Container = styled.section`
+  margin-top: 3.2rem;
+
+  .gender-container {
+    display: flex;
+  }
+
+  div {
+    width: calc(var(--vw, 1vw) * 100 - 5.8rem);
+
+    + div {
+      margin-left: 1rem;
+    }
+  }
+`;
+
+export const Title = styled(Body_4)`
+  color: ${({ theme }) => theme.color.grey_300};
+
+  margin-bottom: 0.8rem;
+`;
