+++ src/pages/information/tags/index.tsx
@@ -1,33 +1,7 @@
 import React from "react";
-import styled from "styled-components";
-import Tag from "../tags/Tag";
-import { ITagsProps } from "../../interfaces/informationPage";
-import { Body_4, Description } from "../../lib/fontStyle";
-
-const Container = styled.section`
-  margin-top: 3.2rem;
-`;
-const Title = styled(Body_4)`
-  color: ${({ theme }) => theme.color.grey_300};
-`;
-const TagContainer = styled.section`
-  display: -webkit-flexbox;
-  display: -ms-flexbox;
-  display: flex;
-  flex-wrap: wrap;
-  -webkit-flex-wrap: wrap;
-  -ms-flex-wrap: wrap;
-`;
-const TagBox = styled.section`
-  margin-top: 0.8rem;
-  margin-right: 0.8rem;
-`;
-const DescriptionBox = styled(Description)`
-  color: ${({ theme }) => theme.color.grey_500};
-
-  font-size: 1.3rem;
-  margin-top: 1.6rem;
-`;
+import Tag from "../../../components/tags/Tag";
+import { ITagsProps } from "../../../interfaces/informationPage";
+import { Container, TagContainer, TagBox, Title, DescriptionBox } from "./index.style";
 
 const Tags = ({ health, setHealth }: ITagsProps) => {
   const handleClick = (e: React.MouseEvent<HTMLDivElement>) => {
